Add standard package
