+++ shardingsphere-error/shardingsphere-common-error/src/main/java/org/apache/shardingsphere/error/SQLExceptionHandler.java
@@ -23,12 +23,12 @@ import org.apache.shardingsphere.error.vendor.VendorError;
 import org.apache.shardingsphere.error.vendor.ShardingSphereVendorError;
 import org.apache.shardingsphere.error.mapper.SQLDialectExceptionMapperFactory;
 import org.apache.shardingsphere.infra.config.exception.ShardingSphereConfigurationException;
-import org.apache.shardingsphere.error.exception.CircuitBreakException;
-import org.apache.shardingsphere.error.exception.ResourceNotExistedException;
-import org.apache.shardingsphere.error.exception.RuleNotExistedException;
-import org.apache.shardingsphere.error.exception.TableLockWaitTimeoutException;
-import org.apache.shardingsphere.error.exception.TableLockedException;
-import org.apache.shardingsphere.error.exception.UnsupportedCommandException;
+import org.apache.shardingsphere.error.exception.standard.CircuitBreakException;
+import org.apache.shardingsphere.error.exception.standard.ResourceNotExistedException;
+import org.apache.shardingsphere.error.exception.standard.RuleNotExistedException;
+import org.apache.shardingsphere.error.exception.standard.TableLockWaitTimeoutException;
+import org.apache.shardingsphere.error.exception.standard.TableLockedException;
+import org.apache.shardingsphere.error.exception.standard.UnsupportedCommandException;
 import org.apache.shardingsphere.infra.util.exception.inside.SQLDialectException;
 import org.apache.shardingsphere.infra.util.exception.inside.ShardingSphereInsideException;
 import org.apache.shardingsphere.sql.parser.exception.SQLParsingException;
