+++ shardingsphere-error/shardingsphere-common-error/src/main/java/org/apache/shardingsphere/error/exception/standard/UnsupportedCommandException.java
@@ -15,24 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.error.exception;
+package org.apache.shardingsphere.error.exception.standard;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.infra.util.exception.inside.ShardingSphereInsideException;
 
 /**
- * Table lock wait timeout exception.
+ * Unsupported command exception.
  */
 @RequiredArgsConstructor
 @Getter
-public final class TableLockWaitTimeoutException extends ShardingSphereInsideException {
+public final class UnsupportedCommandException extends ShardingSphereInsideException {
     
-    private static final long serialVersionUID = 2599713085782288003L;
+    private static final long serialVersionUID = 8010680371699936338L;
     
-    private final String schemaName;
-    
-    private final String tableName;
-    
-    private final long timeoutMilliseconds;
+    private final String commandType;
 }
