+++ shardingsphere-proxy/shardingsphere-proxy-frontend/shardingsphere-proxy-frontend-core/src/test/java/org/apache/shardingsphere/proxy/frontend/exception/UnsupportedCommandExceptionTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.proxy.frontend.exception;
 
-import org.apache.shardingsphere.error.exception.UnsupportedCommandException;
+import org.apache.shardingsphere.error.exception.standard.UnsupportedCommandException;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.is;
